AV-2306 general-ify api resources
